M5Tab_OMIP/pc_software/requirements.txt
pyserial
protobuf
nanopb
Pillow